services/auth/requirements.txt
attrs==22.2.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
exceptiongroup==1.1.0
Flask==2.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.1
tomli==2.0.1
werkzeug==2.0.3
